pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asem-talent-intelligence"
version = "0.1.0"
description = "ASEM-centered semiconductor talent intelligence skeleton for UM Hackathon Domain 2"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "httpx>=0.27,<1.0",
    "pydantic>=2.8,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "pypdfium2>=4.30,<5.0",
    "pypdf>=5.4,<6.0",
    "rapidocr-onnxruntime>=1.2,<2.0",
    "python-docx>=1.1,<2.0",
    "python-multipart>=0.0.9,<1.0",
    "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
